support Linux and 64-bit OS

+++ common/file.h
@@ -13,39 +13,39 @@
 
 typedef struct tagFileDailyData
 {
-    ULONG  ulDate;
-    ULONG  ulBeginWithCheck;
-    ULONG  ulHighWithCheck;
-    ULONG  ulLowWithCheck;
-    ULONG  ulEndWithCheck;
-    ULONG  ulRsv;
-    ULONG  ulVolWithCheck;
+    UINT32  ulDate;
+    UINT32  ulBeginWithCheck;
+    UINT32  ulHighWithCheck;
+    UINT32  ulLowWithCheck;
+    UINT32  ulEndWithCheck;
+    UINT32  ulRsv;
+    UINT32  ulVolWithCheck;
     UCHAR  aucRsv[FILE_THS_DAY_ENTRY_RSV_LEN];
 }FILE_THS_DAILY_ENTRY_S;
 
 typedef struct tagFile5MinuteData
 {
-    ULONG  ulMin;
-    ULONG  ulBeginWithCheck;
-    ULONG  ulHighWithCheck;
-    ULONG  ulLowWithCheck;
-    ULONG  ulEndWithCheck;
-    ULONG  ulRsv;
-    ULONG  ulVolWithCheck;
+    UINT32  ulMin;
+    UINT32  ulBeginWithCheck;
+    UINT32  ulHighWithCheck;
+    UINT32  ulLowWithCheck;
+    UINT32  ulEndWithCheck;
+    UINT32  ulRsv;
+    UINT32  ulVolWithCheck;
     UCHAR  aucRsv[FILE_THS_MIN5_ENTRY_RSV_LEN];
 }FILE_THS_MIN5_ENTRY_S;
 
 typedef struct tagFileWeight
 {
-    ULONG ulQlDate;
-    ULONG ulGift;           //10000 送股
-    ULONG ulOffer;          //10000 配股
-    ULONG ulOfferPrice;     //100   配股价
-    ULONG ulProfit;         //1000  现金红利
-    ULONG ulAdd;            //10000 转增
-    ULONG ulTotalStock;     //总股本
-    ULONG ulCirStock;       //流通股
-    ULONG ulRsv;
+    UINT32 ulQlDate;
+    UINT32 ulGift;           //10000 送股
+    UINT32 ulOffer;          //10000 配股
+    UINT32 ulOfferPrice;     //100   配股价
+    UINT32 ulProfit;         //1000  现金红利
+    UINT32 ulAdd;            //10000 转增
+    UINT32 ulTotalStock;     //总股本
+    UINT32 ulCirStock;       //流通股
+    UINT32 ulRsv;
 }FILE_QL_WEIGHT_ENTRY_S;
 
 ULONG FILE_GetVol(IN ULONG ulFileVol);
